+++ modules/recovery/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+# Version 0.1.1
+
+## Compiler settings
+
+Solidity compiler: [0.8.20](https://github.com/ethereum/solidity/releases/tag/v0.8.20)
+
+Solidity optimizer: enabled via IR with 1.000.000 runs
+
+## Recovery Period
+
+The recovery period is 14 days.
+
+## Expected addresses
+
+- `SocialRecoveryModule` at `0x9BacD92F4687Db306D7ded5d4513a51EA05df25b`
+
+## Changes
+
+### General
+
+- Using the [`safe-singleton-factory`](https://github.com/safe-global/safe-singleton-factory) to deploy contracts.
+
 # Version 0.1.0
 
 ## Compiler settings